refactor(tests): de-barrel tests/e2e/fixtures/ per AGENTS.md
AGENTS.md is unconditional: 'No barrel files (index.ts re-exports).
Import directly from the source file.' tests/CLAUDE.md had a
self-introduced carve-out for tests/e2e/{data,pages,fixtures}/ but
the team's barrel rule doesn't carve those out.

Drop the fixtures barrel and update every spec to import directly
from the fixture file (./fixtures/connectedWallet etc.). `expect`
now comes from @playwright/test directly instead of the re-export.

First of three de-barrel commits (fixtures, then pages, then data).
