+++ tests/e2e/earnPage.spec.ts
@@ -1,7 +1,8 @@
+import { expect } from '@playwright/test';
 import { qase } from 'playwright-qase-reporter';
 
 import { CHAINS } from './data';
-import { expect, connectedTest as test } from './fixtures';
+import { connectedTest as test } from './fixtures/connectedWallet';
 import { EarnPage } from './pages';
 
 import type { Page } from '@playwright/test';
