--- tests/e2e/fixtures/index.ts
@@ -1,4 +1,0 @@
-export { connectedTest } from './connectedWallet';
-export { noWalletTest } from './noWallet';
-export { realWalletTest } from './realWallet';
-export { expect } from '@playwright/test';
