+++ tests/e2e/profileLeaderboard.spec.ts
@@ -1,6 +1,7 @@
+import { expect } from '@playwright/test';
 import { qase } from 'playwright-qase-reporter';
 
-import { expect, connectedTest as test } from './fixtures';
+import { connectedTest as test } from './fixtures/connectedWallet';
 import { MainMenuPage, ProfilePage } from './pages';
 
 test.describe('Profile and Leaderboard navigation with wallet', () => {
